core/requirements.txt
python-dotenv
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=5.26.0
googleapis-common-protos>=1.62.0
